+++ lib/pages/setup.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class Setup extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Center(
+        child: Container(
+          child: Text('Setup Page'),
+        ),
+      ),
+    );
+  }
+}
